Create colorBlobMasks.py
The file contains a routing that uses input threshold file that is passed in command line, to filter image from camera. The blobs of the filtered image are returned as list of tuple. The tuple contains binary mask image, and side of the blob. The size of the blob can later be used to diterming which blob to process first by sorting the blobs by size.