Create doThresholding.py
This script should be called to record thresholds from camera. Then the mouse clicks in window to point the colors.
The file asks name of the file in which the thresholds must be saved. the process repeats so that you can do multiple color thresholding in one run, till you press want to exit.