Update colorBlobMasks.py
Changes the initial code startegy. The list of mask-tupple is sorted to get largest mask on index 0 of the list, increasing probability of match in search that will be done in future.

The largest Blue contour is used to rotate one of the displayed image window compensating for axial rotation of bird.